requirements.txt
django==1.8
pymongo==2.8
django-rest-framework-mongoengine==2.0.2
mongoengine>=0.8.7
djangorestframework>=3.0.0